Switch to Rust edition 2024 and bump goblin to 0.10
Edition 2024 enables if-let chains, which clippy uses to collapse
nested if/if-let blocks across 20+ files (all mechanical, via
cargo clippy --fix).

goblin 0.10 changes pe.image_base from usize to u64, removing one
now-redundant cast in pe.rs. No other API changes affect symdis.

+++ src/binary/pe.rs
@@ -5,7 +5,7 @@
 use std::collections::HashMap;
 use std::path::Path;
 
-use anyhow::{bail, Context, Result};
+use anyhow::{Context, Result, bail};
 use goblin::pe::PE;
 
 use super::{BinaryFile, CpuArch};
@@ -91,7 +91,7 @@ impl PeFile {
             pdata_entries.sort_unstable_by_key(|&(begin, _)| begin);
         }
 
-        let image_base = pe.image_base as u64;
+        let image_base = pe.image_base;
 
         Ok(Self {
             data,
